Update README.md
Panda is a versatile library that simplifies tasks of data manipulation in Python. It is built on the NumPy library and is thus well-suited for working with tabular data and spreadsheets. For this project I aimed to display 2 key functions of the pandas package which are; data wrangling and data visualization using the popular fifa.csv.